refactor: Move range calculation to new file
