+++ packages/core/src/virtualizer/calculateRange.ts
@@ -1,0 +1,239 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import { ScrollDirection } from '../structs/ScrollDirection';
+import type { Offset } from '../types/Offset';
+import type { Rect } from '../types/Rect';
+import { findNearest } from '../utils/findNearest';
+import type { VirtualItem } from './VirtualItem';
+
+const ZERO_RECT: Rect = {
+    height: 0,
+    width: 0,
+};
+
+export const calculateRange = (
+    scrollDirection: ScrollDirection,
+    measurements: VirtualItem[],
+    outerSize: Rect,
+    scrollOffset: Offset
+): {
+    start: number;
+    end: number;
+    maxVisbilityItem: number;
+    visibilities: Record<string, number>;
+} => {
+    let currentOffset = 0;
+
+    switch (scrollDirection) {
+        case ScrollDirection.Horizontal:
+            currentOffset = scrollOffset.left;
+            break;
+        case ScrollDirection.Vertical:
+        default:
+            currentOffset = scrollOffset.top;
+            break;
+    }
+
+    const size = measurements.length - 1;
+    const getOffset = (index: number) => {
+        switch (scrollDirection) {
+            case ScrollDirection.Horizontal:
+                return measurements[index].start.left;
+            case ScrollDirection.Both:
+            case ScrollDirection.Vertical:
+            default:
+                return measurements[index].start.top;
+        }
+    };
+
+    let start = findNearest(0, size, currentOffset, getOffset);
+    if (scrollDirection === ScrollDirection.Both) {
+        // Find the first item in the same row which has the same `top` value
+        // and the `left` value is greater than `scrollOffset.left`
+        const startTop = measurements[start].start.top;
+        while (
+            start - 1 >= 0 &&
+            measurements[start - 1].start.top === startTop &&
+            measurements[start - 1].start.left >= scrollOffset.left
+        ) {
+            start--;
+        }
+    }
+
+    let end = start;
+    // The visiblities of each item
+    const visibilities: Record<string, number> = {};
+    let maxVisbilityItem = start;
+    let maxVisbility = -1;
+    while (end <= size) {
+        const itemRect = measurements[end].size;
+        const visibility = ZERO_RECT;
+        const topLeftCorner = {
+            top: measurements[end].start.top - scrollOffset.top,
+            left: measurements[end].start.left - scrollOffset.left,
+        };
+        const visibleSize = {
+            height: outerSize.height - topLeftCorner.top,
+            width: outerSize.width - topLeftCorner.left,
+        };
+
+        //          |                        |
+        //    (8)   |           (1)          |   (2)
+        //          |                        |
+        // ─ ─ ─ ─ ─┼────────────────────────┼─ ─ ─ ─ ─
+        //          |                        |
+        //          |                        |
+        //    (7)   |      Container         |   (3)
+        //          |                        |
+        //          |                        |
+        //          |                        |
+        // ─ ─ ─ ─ ─┼────────────────────────┼─ ─ ─ ─ ─
+        //          |                        |
+        //    (6)   |           (5)          |   (4)
+        //          |                        |
+        if (scrollDirection === ScrollDirection.Horizontal && visibleSize.width < 0) {
+            // The top left corner belongs to the (2, 3, 4) areas
+            end--;
+            break;
+        }
+        if (scrollDirection === ScrollDirection.Vertical && visibleSize.height < 0) {
+            // The top left corner belongs to the (4, 5) areas
+            end--;
+            break;
+        }
+        if (scrollDirection === ScrollDirection.Both && (visibleSize.width < 0 || visibleSize.height < 0)) {
+            // The top left corner belongs to the (2, 3, 4, 5) areas
+            end--;
+            break;
+        }
+
+        // Calculate the width that is visible within the container
+        // We don't care it in the vertical scroll mode
+        if (scrollDirection === ScrollDirection.Vertical) {
+            visibility.width = 1;
+        } else if (topLeftCorner.left < 0) {
+            // The top left corner belongs to the (1) area
+            // 1) The top right corner is inside of the container
+            // |   (top, left)          |
+            // |        ●─ ─ ─ ─ ─●     |   visibility.width = 1
+            // |           (top, right) |
+            // |                        |
+            // ┌────────────────────────┐
+            // |                        |
+            // |                        |
+            // |                        |
+            // |                        |
+            // |                        |
+            // └────────────────────────┘
+
+            // 2) The top right corner is outside of the container
+            // |   (top, left)          |       (top, right)
+            // |        ●─ ─ ─ ─ ─ ─ ─ ─┼─ ─ ─ ─ ─ ─●
+            // |                        |
+            // ┌────────────────────────┐
+            // |                        |
+            // |                        |
+            // |                        |
+            // |                        |
+            // |                        |
+            // └────────────────────────┘
+            const visibleWidth = itemRect.width - -topLeftCorner.left;
+            visibility.width = visibleWidth <= outerSize.width ? visibleWidth / outerSize.width : 1;
+        } else {
+            // The top left corner is inside of the container
+            // 1) The top right corner is inside of the container
+            // ┌────────────────────────┐
+            // |   (top, left)          |
+            // |        ●─ ─ ─ ─ ─●     |   visibility.width = 1
+            // |           (top, right) |
+            // |                        |
+            // |                        |
+            // └────────────────────────┘
+
+            // 2) The top right corner is outside of the container
+            // ┌────────────────────────┐
+            // |   (top, left)          |       (top, right)
+            // |        ●─ ─ ─ ─ ─ ─ ─ ─┼─ ─ ─ ─ ─ ─●
+            // |                        |
+            // |                        |   visibility.width = visibleSize.width / itemRect.width;
+            // |                        |
+            // └────────────────────────┘
+            visibility.width = itemRect.width <= visibleSize.width ? 1 : visibleSize.width / itemRect.width;
+        }
+
+        // Calculate the height that is visible within the container
+        // There are four cases
+        if (scrollDirection === ScrollDirection.Horizontal) {
+            visibility.height = 1;
+        } else if (topLeftCorner.top < 0) {
+            // The top left corner belongs to the (1) area
+            // 1) The bottom left corner is inside of the container
+            // |   (top, left)          |
+            // |        ●               |
+            // |        |               |
+            // ┌────────┼───────────────┐   visibility.height = visibleHeight / outerSize.height;
+            // |        |               |
+            // |        |               |
+            // |        ● (bottom, left)|
+            // |                        |
+            // └────────────────────────┘
+
+            // 2) The bottom left corner is outside of the container
+            // |   (top, left)          |
+            // |        ●               |
+            // |        |               |
+            // ┌────────┼───────────────┐
+            // |        |               |   visibility.height = 1;
+            // |        |               |
+            // |        |               |
+            // |        |               |
+            // └────────┼───────────────┘
+            // |        |               |
+            // |        ● (bottom, left)|
+            const visibleHeight = itemRect.height - -topLeftCorner.top;
+            visibility.height = visibleHeight <= outerSize.height ? visibleHeight / outerSize.height : 1;
+        } else {
+            // The top left corner is inside of the container
+            // 1) The bottom left corner is inside of the container
+            // ┌────────────────────────┐
+            // |   (top, left)          |
+            // |        ●               |
+            // |        |               |   visibility.height = 1;
+            // |        |               |
+            // |        ● (bottom, left)|
+            // └────────────────────────┘
+
+            // 2) The bottom left corner is outside of the container
+            // ┌────────────────────────┐
+            // |   (top, left)          |
+            // |        ●               |
+            // |        |               |   visibility.height = visibleSize.height / itemRect.height;
+            // |        |               |
+            // |        |               |
+            // └────────┼───────────────┘
+            //          |
+            //          ● (bottom, left)
+            visibility.height = itemRect.height <= visibleSize.height ? 1 : visibleSize.height / itemRect.height;
+        }
+
+        visibilities[end] = visibility.width * visibility.height;
+        if (maxVisbility < visibilities[end]) {
+            maxVisbility = visibilities[end];
+            maxVisbilityItem = end;
+        }
+        end++;
+    }
+
+    return {
+        start,
+        end,
+        maxVisbilityItem,
+        visibilities,
+    };
+};
